Purge prerendered pages and cache on deploy
Otherwise, the cached HTML from a previous release might continue to be
served after a new release, causing 404s when we try to serve old JS and
CSS assets (referenced from the cached prerendered HTML).

In addition to clearing the cache, we also need to delete the
prerendered HTML file, so that the cache can't be repopulated by a
request that happens to come in between the time of the deploy and the
time that a new prerendered page is uploaded to S3.