Remove `thread_safe` gem
It used to be necessary to have this gem in our `Gemfile` because
`active_model_serializers` depended on it, but didn't list it in its
gemspec. `active_model_serializers` seems to no longer depend on
`thread_safe`, so let's remove it.

Credit to https://github.com/BigBigDoudou/gems_bond for helping to bring
this to my attention.