Increase Ferrum timeout from 10s to 60s when prerendering
When a bunch of deploys are going out ~simultaneously, and thus
triggering dyno restarts, it can happen that a request won't be served
within 10 seconds. Our web dynos seem to take almost ~40 seconds to boot
up. Therefore, this change increases the Ferrum timeout to 60 seconds
(giving ~20+ seconds of "padding", even if a prerender request is
initiated at the same time as a dyno restart).