Don't use async partials for quiz#show after submitting an answer
The async partials are mostly useful when advancing to the next
question, because then everyone needs to load the quiz question at once.
When a user submits their answer, though, they are probably the only
person submitting at that time, so there's less need for the async
partials, and not using the async partials provides a better user
experience (by not rendering the loading placeholders for a while).