Bump `heroku-deploy` from 3.6.8 to 3.12.12
We are getting an error (see below) when the workflow attempts to deploy
to Heroku, so hopefully this will fix it. :shrug:

```
Error: Command failed: git push heroku HEAD:refs/heads/master --force
fatal: 'heroku' does not appear to be a git repository
fatal: Could not read from remote repository.
```