Store asset size timeseries in Postgres rather than Redis
Storing them in Redis was using too much memory when reading the values
because the timestamps and values were stored as strings. By storing
them in a Postgres JSON array, we can leave them as integers and floats,
saving memory when reading the values. I think inserts will be a bit
slower and use more memory, but that's okay.