Bundle admin JS separately
When admin JS is bundled with the main application JS, it can make the
application JS have a heavier page weight and/or be split into more
bundles than when admin JS is bundled separately. In particular, we are
seeing this with how chartkick requires all of the chart.js library, but
the logs app doesn't. Bundled together, all of chart.js is served to
logs users. When bundled separately, we can serve just the needed
chart.js functionality to them (though that will also require better
tree shaking).