Have HTML controller tests depend on (wait for) feature tests
The HTML controller tests don't actually need the feature tests to run
first. We are doing this simply so that the feature tests will run
first. This is desirable because the feature tests include Percy checks,
which take time, so it's better to start the feature tests sooner rather
than later.