Add back `pry-byebug` gem
We removed this gem in f8f3037 / #1232 because it was printing a
deprecation warning. I think that has since been fixed. I'm adding this
gem back to the development and test bundle groups because I think it
has different/better debugging behaviour sometimes in multi-threaded
environments (e.g. in capybara specs when we need the server thread to
be executed so we can navigate to new pages and make API requests, see
https://github.com/deivid-rodriguez/pry-byebug/pull/ 142 ).