Remove debug gem
We switched back to `pry-byebug` in 3a7ce8e (see more details there). As
a result, we aren't really using `debug` anymore. Much love to the gem,
and I think it's great to have a debugging tool in Ruby core, and it
works pretty well, but it's not for me right now.