Remove InvalidRecordsCheck workers and mailer
We now have enough records that running these jobs is starting to take a
somewhat noticeable amount of time and use a noticeable amount of
memory. This could be further optimized, but the time/effort to do that
doesn't seem worth it, since this worker pretty much never finds any
invalid records, anyway.