Use pinia (no longer vuex) for Vue/JavaScript state management
Pinia is the officially endorsed Vue state management library now, and I
do think it's a little nicer to use and simpler / more straightforward /
better aligned with my mental model than Vuex. It's a little bit verbose
with respect to always importing the store(s) into each component, and
adding the store(s) to each component's `data`/`setup`, but I think that
overall it's a win.