Use standalone `workoutIsInProgress` attribute (not `workout.active`)
This fixes a bug wherein the user was not shown the new-workout-form
upon visiting `/workout` if they had previously completed a workout. By
storing `active` as a property of `workout`, we were posting that to the
server and saving it in the user's preferences upon workout completion,
which isn't what we want.