Allow enforcing uniqueness of Sidekiq jobs/arguments while executing
When the `unique_while_executing` option is set for a worker class, then
no other instances of that job should execute concurrently with a given
set of arguments while one instance of that worker is executing with
those arguments. If an attempt is made to execute such a job
concurrently, then it will be rescheduled to retry after a random
interval (currently between 1 and 30 seconds later).

This simplifies the `CreateIpBlock` worker (making it no longer
necessary to guard against a certain race condition therein) and
hopefully will also avoid certain errors in the `SaveRequest` worker.