Rename/replace `PerformantImage` component with `LoadNotifyingImage`
Now, the webp image format is supported by all browsers that we care
about, so we no longer need the ability of the `PerformantImage`
compontent to render either a webp image or a fallback type.

However, the functionality of the component wherein it fires an event
upon image load is still useful/needed (so we can recalculate the
homepage waypoints after image load). Therefore, we're trimming down the
compontent to just that scope and renaming it accordingly.