Move Sidekiq deploy marking back into `release-tasks`
The way we were doing it with `bin/_sidekiq/mark_deploy.rb` takes ~2
seconds whereas it takes just milliseconds to do it this way. Modifying
ENV doesn't feel great, but neither does taking 2 seconds to mark the
deploy when we could instead do it in less than 0.1 seconds.

(mperham is apparently uninterested in changing this in Sidekiq so we
don't have to do something like this; see
https://github.com/mperham/sidekiq/issues/ 5615 .)