Update JavaScript packages
`yarn upgrade` plus other stuff

I'm doing this because dependabot didn't update `vite`, even though
there's a new release of `vite` available (4.0.0). This doesn't fix
this, but maybe it will lay the groundwork for a fix.

This bumps us from chart.js 3 to 4.

We're now installing `chartkick` from GitHub because no version has been
released to NPM yet that supports charts.js 4.

We're switching from `date-fns` to `luxon` because I was having a
problem with `date-fns` (an error message about "This method is not
implemented: Check that a complete date adapter is provided.") and
`luxon` seems to produce a smaller bundle size.