[groceries] Update existing item objects after reconnecting
This fixes at least one bug: skipped items becoming unskipped after
reconnecting to the Internet, because the list of skipped items had a
reference to the old item object(s), but the stores were previously
being updated with new item objects.

Unfortunately, it's hard to write a feature test for this, because I
haven't figured out how to disconnect from the Internet (affecting a
WebSocket connection) and reconnect with capybara. JavaScript unit tests
might answer, but that'd probably be a pretty annoying test to write,
and we don't have a JavaScript unit test harness set up right now.