Downgrade `sidekiq` from 7.0.7 to 7.0.6
fixes rb#562

https://rollbar.com/davidjrunger/davidrunger/items/562/

I'm not sure if this will actually fix the above exception from
occurring, but I think it's worth a shot. I think that the exception in
question appeared within a day after deploying a change that bumped
Sidekiq from 7.0.6 to 7.0.7, and that exception is occurring with a
Sidekiq-only (plus redis dependencies) stack trace, and also the Sidekiq
changelog for 7.0.7 includes "Fix redis-client API usage which could
result in stuck Redis connections" which sounds potentially related, so
let's see whether this downgrade will fix the issue.