Have Selenium devtools version fallback to previous version if needed
When a new version of Chrome is released, then specs fail with the error
below until we can bump our `selenium-devtools` gem to match the newly
released Chrome version, which is especially problematic because a new
version of the gem is not always released by the time that the new
Chrome version is released.

```
LoadError:
  cannot load such file -- selenium/devtools/v112
```

With this change, instead of erroring, we will fallback to using the
previous devtools version from the `selenium-devtools` gem, which should
hopefully be available.