Switch from yarn to pnpm
Yarn is really on my bad side. Apparently for yarn version 3 I'm
supposed to commit the actual yarn executable to my git repo? You've
gotta be kidding me. So the alternative is to stick with yarn version 1
which I guess isn't getting updated anymore, which is also bad.

pnpm seems to get great reviews on /r/javascript, so let's try it.