Exclude gitignored files and schema.rb from RuboCop
This overrides the `runger_style` Excludes list. The benefit of this is
that we can simply run `bin/rubocop` and it will lint only the files in
the repo that we want to be linted (e.g. not Ruby files in `personal/`).
I started to explore this because I thought that running RuboCop against
the whole repo (rather than specific files) might be faster by better
leveraging the prism parser. I think that might not be true, but this
nevertheless seems like a good change, although the downside is that now
`git` must be available, and also executing that `git` command takes
some time.

This cannot be done in the `runger_style` config, because then the
working directory for the `git status --ignored --porcelain` command is
`runger_style` rather than the project repo.