Exclude gitignored files and schema.rb from RuboCop (#4009)
This overrides the `runger_style` Excludes list. The benefit of this is
that we can simply run `bin/rubocop` and it will lint only the files in
the repo that we want to be linted (e.g. not Ruby files in `personal/`).

I started to explore this because I think that running RuboCop against
the whole repo, rather than a list of specific spec files, can be a lot
faster.

This cannot be done in the `runger_style` config, because then the
working directory for the `git status --ignored --porcelain` command is
`runger_style` rather than the project repo.