Sleep a half second after devicon font loads for Percy snapshot
A [Percy build][1] flaked recently (maybe somehow due to the switch to
Tailwind? the timing seems very suspicious) so I am going to add this
extra half second sleep to try to allow enough time for fonts/CSS to be
parsed etc for a stable Percy snapshot of the homepage.

1: https://percy.io/David-Runger/david_runger/builds/34790045/changed/1902141723