Store item skipp as boolean on item object (not skippedItems array) (#4408)
I think it's easier to think about this way, the code is simpler / more
concise, and it should also be more performant. I'm not sure why I used
a skippedItems array initially. I'm guessing because it seemed wrong to
put "ephemeral"/"optional" state on the Item type (which, up until now,
has contained only required properties and properties that come from the
database), but I don't think that this is really a compelling reason.