Import chartjs-adapter-luxon after all other imports (#4463)
Because `chartjs-adapter-luxon` is an import used for its "side
effects", `@ianvs/prettier-plugin-sort-imports` cannot be confident in
moving it farther up. Therefore, in preparation for our upcoming mega
Prettier formatting PR, we'll instead move it all the way to the bottom,
where it will be left by `@ianvs/prettier-plugin-sort-imports`, as
opposed to being left randomly in the middle of the import list.