Remove base element styles from text_log.vue component (#4465)
Move the heading styles that are not already in tailwind_overrides.scss
there.

The `strong` and `em` tags seem to already be styled as desired by
browsers and/or the Tailwind reset.

Motivations: Simplify the code. Reduce wasted CSS page weight. These
styles would be formatted by Prettier in a way that is unacceptable to
stylelint (no newlines between the rules).