Replace lefthook with executable git hook script(s) (#4478)
We added our lefthook config in #1693 / 8004fc89, and this removes it.

Instead, we are now going to manually manage git hooks as executable
scripts that I define. I used this article as a reference:
https://www.viget.com/articles/two-ways-to-share-git-hooks-with-your-team/
.

I'm adding a check in an initializer to make it more likely that persons
developing on this repo will set up the git hooks.