[home] Make davidrunger app links relative & use path helpers (#4509)
One motivation for this change is that, in development, clicking these
links will go to the localhost version of the page, which I think is
usually what we want. Also, including the domain name in the links is
not very DRY. Also, using the path helpers should more quickly alert us
if we remove/change one of these routes.