Add csv gem (being extracted from stdlib) to Gemfile [DEV-7] (#4510)
Requiring the `csv` gem via the Gemfile / Gemfile.lock will be necessary
as of Ruby 3.4. In our case, we are already somewhat covered by the fact
that ActiveAdmin now has `csv` as a dependency, but we should also add
it directly ourselves, since our app uses `csv` directly (and we might
someday remove ActiveAdmin).