Run non-Gitleaks pre-push checks in the background [DEV-27]
This will get branches up to GitHub more quickly, and also _might_ be
overall somewhat faster, since this will allow for some parallelization
of the linting.

We're still going to run Gitleaks in a push-blocking serial fashion
because we don't want it to be possible to push a secret to GitHub even
for a moment.