Run non-Gitleaks pre-push checks in the background [DEV-27] (#4525)
This will get branches up to GitHub more quickly, and also _might_ be
overall somewhat faster, since this will allow for some parallelization
of the linting.

We're still going to run Gitleaks in a push-blocking serial fashion
because we don't want it to be possible to push a secret to GitHub even
for a moment.

See
https://github.com/davidrunger/dotfiles/commit/71f7c24ccd97e3099ce6fedf05885c81d7d63e0a
, which added to my dotfiles repo the necessary supporting scripts that
are referenced herein (e.g. `background-and-notify` and
`write-background-notification`).