Remove @formkit/auto-animate [LOG-1]
The library did what it promised, but I think it really hurts
performance on a text log page with a lot of entries, so I am going to
remove it.

It might be nice, in the future, to manually animate the appearance of a
new text log entry into the list. I've created an issue for that:
https://davidrunger.atlassian.net/browse/LOG-9 .