[pre-push] Remove TypeScript check (#4565)
Motivation: The TypeScript check seems to use up a pretty nontrivial
amount of the not overly abundant resources on my little computer. And I
think it's relatively rare that I push a PR that has TypeScript
violations. So I'm not sure it's worth the cost (somewhat overwhelming
my computer) to run the TypeScript checks locally. We will still check
types in CI.