Rescue Ferrum::ProcessTimeoutError and retry in Prerenderer
Sometimes we get errors (Browser did not produce websocket url within 20
seconds, try to increase `:process_timeout`). Rather than increasing the
process timeout, as suggested in that error message, I'm going to try
retrying, which will effectively double the total amount of time we
spend waiting, and I wonder if retrying might be more reliable than just
waiting longer (in case maybe something went irrecoverably wrong with
the first initialization attempt).

I'm going to set a timeout for the overall deploy-and-prerender GitHub
action, in case this keeps retrying endlessly, which I think is at least
a theoretical risk. I'm also going to tighten up the test timeout, while
I'm at it. We can increase that, in the future, if we start bumping up
against it, but I don't think that we currently ever do.