[groceries/check-in] Keep bottom buttons always in view (#4596)
Prior to this change, if there were a lot of items to check in, then the
"Cancel" and "Check in items in cart" buttons would be initially
scrolled out of view and one would have to scroll down in order to
interact with them. With this change, it's never necessary to scroll in
order to interact with these buttons; they are kept always in view.