Remove Babel skill row and devIconStyle prop
I'm not much of an expert in Babel, and I don't interact with it much
directly these days (if I ever did). Instead, at this point, if I'm
using Babel at all (which I might not even be), it seems like Babel is
managed by my tools.

Since the Babel skill row was the only one using the `devIconStyle`
prop, we can remove it.