[groceries] Animate new items using Vue TransitionGroup [GROC-11] (#4612)
One advantage of this is that we no longer need to manually manage a
`newlyAdded` property on the grocery items (which always felt a little
hacky). Another advantage is that this actually smoothly animates the
other items up and down, I think, whereas previously the other items
would jerk up and down.