Enable YJIT via an initializer (#4630)
This is better than doing it via an env var, because this way we don't
JIT the boot time stuff, which apparently is a good thing.
https://youtu.be/--0pXVadtII?si=BEC8-FnDyJw5nIED&t=748

Another thing I like about this change is that it puts the config in the
codebase, rather than in an environment variable. I don't think that
this setting is something that we will want to be changing/toggling on
short notice (which would make an environment variable relatively
compelling).

I have removed the `RUBY_YJIT_ENABLE=1` env var from the Dokku config.