Downgrade tailwindcss from 3.4.6 to 3.4.4 (#4637)
A bug seems to have been introduced, I think in 3.4.5, which is
currently causing some visual regressions in production. The problem
seems to be that, in some cases, when using pug templates, the last
listed class does not get picked up (for example, a `z-10` class in
Modal.vue and `leading-unset` in Item.vue).

Possibly relevant issues:

https://github.com/tailwindlabs/tailwindcss/issues/ 14005

https://github.com/tailwindlabs/tailwindcss/issues/ 14013