[workout] Migrate all components to script setup [DEV-44]
Also, rename the directory from "workouts" to "workout" (to reflect the
canonical route) and also move all non-main components into a
`components/` subdirectory (which is a pattern that we use for all other
Vue apps, which I think is nice, to keep things a bit less cluttered at
the top level of the directory).

Also, enable the `object-shorthand` ESLint rule, which I guess is not
enabled by default, at least not in our config.