[groceries] Extract a CheckInItem component [GROC-14] (#4672)
I ended up needing to move some of the functionality to the store
because of a rule against mutating a prop. I guess that this is probably
a good change. We were getting away with it before, I guess, because the
prop was `items` (an array) and we were mutating its elements directly,
which I guess wasn't detected as a violation.