Use freezolite (no more frozen_string_literal comments) [DEV-2] (#4679)
One downside of this is that, because we no longer have explicit
`frozen_string_literal: true` comments at the top of files, RuboCop does
not know that strings are frozen by default, and so RuboCop's
autocorrections (applied in this change) include adding explicit
`.freeze` calls to string constants. This is not totally ideal, but I
think it's okay.

I wonder if maybe I am going to miss having these comments at the top of
my files. I sort of felt like they provided a pleasant bit of "breathing
room" at the top of the file, rather than smashing the top of the class
definition in a file up against the top of my editor.

I guess let's try it and see how it feels.