Replace "\r\r\n" with "\r\n" when parsing incoming emails [LOG-2]
I'm not totally sure if these character sequences occur in production,
but they were occurring in my development testing, and I do think that
they probably occur in production, too.

Prior to this change, I was able to reproduce the bug. With this change,
the bug no longer occurs. So, I think this has a decent chance of fixing
the bug in production, too.