Add a dedicated proposal page (#4694)
GROC~3 calls for us to link to a place where users can invite their
partner to the app as their partner. The only place that we currently
have to do that has a URL of `/check_ins`. We want to add a dedicated
place with a more appropriate URL as a generic place to make a proposal.