[logs] Only focus logInput if it exists
I saw an error in my JS console once about "cannot focus undefined" or
whatnot (while working on LOG~5). I think TypeScript was trying to warn
me about this, too; this change lets us remove `as unknown`.

I think that this might reflect a race condition?, where if we try to
focus the logInput, but it doesn't yet exist and only later gets
rendered, then the logInput will not be automatically focused? If so,
that's not ideal, but I guess it's no worse than the current situation,
where the same thing will happen, and also a JavaScript exception will
be thrown.