Don't assume that template refs exist (#4717)
I saw an error in my JS console (while working on LOG~5) about "cannot
focus undefined" or whatnot. I think TypeScript was trying to warn me
about this, too; this change lets us remove `as unknown`.

I think that this might reflect a race condition?, where if we try to
focus the logInput, but it doesn't yet exist and only later gets
rendered, then the logInput will never be automatically focused? If so,
then that's not ideal, but I guess it's no worse than the current
situation, where the same thing will happen, and also a JavaScript
exception will be thrown. So, this is probably at least an improvement.
This conditional also, I think, more explicitly suggests that potential
issue, which is also probably a good thing. Maybe we'll fix it, in the
future (if, indeed, it's a real risk).