Lazy-load offscreen images [HOME-3]
This is easy to do by simply using browsers' native `loading='lazy'`
attribute for the `img`s.

I this change, we're adding this attribute to four images. Chrome
suggested three of them, but I also decided to add it to `david.jpg`, as
well, even though Chrome did not suggest that (I'm guessing because
Chrome is not going to lazy-load it, anyway, since it's near enough to
the top of the page).

NOTE: Due to a Firefox bug, the `loading` attribute must come _before_
the `src` attribute. https://stackoverflow.com/a/76252772/4009384 /
https://bugzilla.mozilla.org/show_bug.cgi?id=1647077 .