Store Google OAuth secrets in ENV vars (not Rails credentials)
I recently lost the encryption secret for the development credentials
file (which is why I'm also deleting the encrypted file in this commit).
Rather than rebuild a new credentials file (which sort of demonstrates
how credentials can be a bit of a pain), instead, let's go back to just
using ENV vars for this.

I have just set the appropriate ENV vars in production, so this should
be a smooth, no-downtime deployment.