Delete encrypted Google credentials (#4732)
Just creating this PR was a bit of a pain, in some ways, making me think
that I made the right decision to start moving away from Rails
credentials. For example, to edit the production credentials, I needed
to comment out the lines in `config/environments/production.rb` that
require a `MEMCACHED_URL` etc. And some of my tooling ran into issues
trying to decrypt the diff, and, when it worked, that process was a bit
slow.