Deploy to Dokku using Dockerfile with jemalloc [DEV-13][DEV-14]
By deleting the `.buildpacks` file (and by already having a
`Dockerfile`), Dokku is now supposedly going to use the Dockerfile
during the deployment process.

Most of the Dockerfile modifications have been copied from the
Dockerfile that was generated for an app created on 2024-07-28 using the
Rails `main` branch, though I have made some modifications to that
template.

In the production deployment, Dokku sets up the database services and
coordinates them via Docker. In lieu of running Dokku locally, though, I
am also modifying the `docker-compose.yml` file in this commit. This
file won't be used in production by Dokku, but rather is something that
I can use locally to run database services to which the Docker web
container can connect (in order to sanity check locally that things are
working).

I'm deleting `docker-vite.sh` and removing the `vite` section of the
`docker-compose.yml` because the previous version of our `Dockerfile`
had bene built for development purposes, whereas this version is being
built for production purposes (and we will download precompiled assets,
rather than building them using Node/Vite via the Dockerfile).

**Build:** `rmrf public/assets/ public/vite/ public/vite-admin && docker compose build --build-arg GIT_REV=$(git log origin/main --format=format:%H | head -n 1) --build-arg RAILS_MASTER_KEY=$(cat config/credentials/production.key) --build-arg RUBY_VERSION=$(cat .ruby-version) --progress=plain`

**Run server:** `docker compose run --service-ports -e "GIT_REV=$(git log origin/main --format=format:%H | head -n 1)" -e "RAILS_MASTER_KEY=$(cat config/credentials/production.key)" rails bin/rails server`

**Run Sidekiq:** `docker compose run -e "GIT_REV=$(git log origin/main --format=format:%H | head -n 1)" -e "RAILS_MASTER_KEY=$(cat config/credentials/production.key)" rails bin/sidekiq`

**Skedjewel / Release tasks:** copy command from `Procfile` and replace
in the above Sidekiq command.