Stop installing pkg-config
We don't seem to need it.

We do ned `libpq-dev`. Removing it causes `pg` installation to fail.

I think we'll need `git` if we want to be able to install packages from
a git repository. Even though we aren't currently doing this, I think
it's worth just adding support for it, in case we do in the future.

I didn't try removing it, but I'm guessing we do need `build-essential`,
as well, to be able to compile C extensions.