Add dokku predeploy script (to run assets:precompile) to app.json
I think that this was previously done for us automatically as part of
the Heroku Ruby buildpack, but, now that we aren't using that anymore, I
think that we need to specify it manually. Thanks to
https://pawelurbanek.com/optimize-dokku-deployment-speed for this
callout/tip (and, in general, some really useful information and
perspective about this process).