Set ARG GEMS_DIRECTORY higher up
I think that, when set in the context of the build phase, it wasn't
available outside of that phase (based on the relevant COPY step logging
as just `COPY --from=build` without saying the directory name(s)).
Hopefully this will make the ARG (which we're really just using as a
variable) available in all of the places that we need it.