[Dockerfile] Remove RAILS_MASTER_KEY build arg
Motivation: make the build process slightly simpler.

Instead of having this as a build arg, we'll expect the production
RAILS_MASTER_KEY to be present in a `.env` file on the machine doing the
building.

Also, just put DOCKER_BUILD and GIT_REV as env variable prefixes for the
assets:precompile command that needs them, rather than adding them, in
general, to the ENV.