[docker-compose] Add most Rails services to the external network
When a service is only part of the internal network, I think that it
cannot even access the outside Internet. We still want that for some
services (i.e. our databases, which have no need to communicate with
the Internet), but most of our Rails services need to be able to
communicate out (e.g. to download assets from S3 or to send errors to
Rollbar). I think that `clock` does not need access to the external
Internet, though, so I'm just putting that Rails service on the
`internal` network only.