[docker-compose] Add most Rails services to the external network (#4776)
When a service is only part of the internal network, I think that it
cannot even access the outside Internet. We do want that behavior for
some services (i.e. our databases, which have no need to communicate
with the Internet), but most of our Rails services need to be able to
communicate out (e.g. to download assets from S3 or to send errors to
Rollbar). I think that `clock` does not need access to the external
Internet, though, so I'm just putting that Rails service on the
`internal` network only.

The services that we are adding in this change to the `external` network
still should not have their ports exposed to the external internet,
since we aren't using a `--ports` CLI flag or a `ports` option in
`docker-compose.yml` (except for `nginx`, which we do want/need to
expose to the external Internet).