[routes] Specify Content-Type text/plain for routes responses (#4784)
Currently, when going to `/google83c07e1014ea4a70` or `/sha`, my browser
downloads the content. I think that this is because of the `add_header
X-Content-Type-Options "nosniff" always;` option in our NGINX config. I
think that adding this Content-Type header will allow the browser to
display the content directly in the browser, as it would do before, and
as we want.

Also, show the whole git SHA, rather than just the first 8 characters.
The git SHAs are all available on the public GitHub, anyway, and I think
being able to see the full SHA is nice.