Delete Procfile
Instead, the services that we run is now managed by the `docker compose
up` line in our deploy.sh script and the specific definitions in
`docker-compose.yml` of the services that are booted by that command.

One thing to note is the `-p $PORT` for the Rails server. That is
something that we might want to bring back, in some form, in the future.
Currently, we always run the server on port 3000. The problem is that
this doesn't allow booting multiple servers simultaneously for a
zero-downtime rolling deployment. So, like I say, we might bring back
some form of this flexible port option.