Don't use rails runner trick to execute release-tasks script (#4799)
I think that the `exec` in here might be causing the remaining bits of
the `deploy.sh` script not to exit; I think that when `release-tasks`
exits, the whole process exits. At least, that would explain what we're
seeing in the logs:
https://github.com/davidrunger/david_runger/actions/runs/10225938665/job/28295635507
and on the server where `main` is not getting checked out and updated at
the end of the deployment process.