[deploy] Run the server install script as part of each deploy (#4805)
Usually, this will have no effect (since the install script will usually
have already been executed, and since it should be idempotent), but
running the install script each time that we deploy makes it easy to
ensure that the latest version is always "installed". Since the install
script should be fast and idempotent, there shouldn't really be much
downside to doing this.