[docker] Make RAILS_ENV a build arg [DEV-70] (#4815)
This way, we can build with `RAILS_ENV=production` for production and
`RAILS_ENV=development` when running locally.

However, even when we boot the app locally with `RAILS_ENV=development`,
we still don't install gems from the `development` or `test` bundler
groups. So, this PR adds a number of additional checks against a new
`IS_DOCKER` constant, so that we don't try to use development-/test-
only gems when running the app locally via Docker (with
`RAILS_ENV=development`).