[docker-compose] Rework depends_on listings for Rails services (#4820)
The problem with what we were doing before is that we had some
service-specific `depends_on` declarations that were overriding (as
opposed to adding to) the default `depends_on` declaration for
`x-rails-config`. For example, `clock` was not depending on `redis`,
which it should do.

I think that this might be the cause of errors like this when trying to
boot `clock`:

> clock-1 | Skedjewel 0.0.13 is running with PID 1. > clock-1 |
Unhandled exception: Socket::ConnectError: Error connecting to
'localhost:6379': Connection refused (Redis::CannotConnectError)

This new configuration has most Rails services `depends_on`
initialize_database. The exception is initialize_database itself, which
depends on memcached, postgres, and redis. Hopefully this setup will
ensure a reliable bootup chain.