[nginx] Remove 'Forwarded' header
We noticed that rack-attack was blocking the Cloudflare IP that
forwarded the bad request rather than the originating client IP.

I figured out a fix for this: #4823. However, I have been wondering why
that fix was even necessary.

I suspect that it might have been necessary because of the fact that
[Rack prioritizes][1] the `Forwarded` header over `X-Forwarded-For`. I
suspect that maybe the way that these headers are handled (maybe as they
relate to `cloudflare-rails`?) is different, and/or maybe we aren't
formatting the `Forwarded` header correctly. Maybe on Dokku (before our
recent migration to raw Docker (Compose)) we weren't sending a
`Forwarded` header at all, and that's why we didn't need #4823 when on
Dokku.

[1]: https://github.com/rack/rack/blob/v3.1.7/lib/rack/request.rb#L43

Therefore, I am going to try reverting the #4823 change and also stop
sending a `Forwarded` header. If, after this change, rack-attack
continues to block the originating client IP rather than the Cloudflare
IP, then my above theory is probably correct, and we'll leave things as
they will at that point be. If we see a regression, then my theory is
probably not correct, and we'll probably revert this PR. But I would
like to try it and see what happens.

Unfortunately, it's sort of necessary to test this in production,
because outside of a production environment it is difficult to simulate
what the real request headers and environment etc will be when these
requests are handled.

One nice thing about this change is that it allows us to delete a decent
amount of NGINX code.