[deploy] Add '--remove-orphans' to 'docker compose up' command (#4827)
This should hopefully prevent disk usage from growing whenever we
deploy, and maybe eventually we will be able to do away with the `docker
system prune --all --force` cron job.

The reason I initially hesitated to do this is because I didn't want to
waste time/CPU deleting orphans when trying to boot up new containers.
However, from using the `--remove-orphans` option locally, it doesn't
feel like it adds a lot of time, and so I think it might be worth doing.