[docker-compose/rails] Remove hardcoded 'production' RAILS_ENV
In #4815, we intended to make it possible to build the Docker app in
development mode or production mode.

However, due to the hardcoded RAILS_ENV being removed in this PR, we
actually were still booting the app in a production Rails environment.

This change makes it so that when we build the app for development
(`bin/build-docker development`) and run it with Docker Compose (`docker
compose up --detach --remove-orphans`), the app will run with
`RAILS_ENV=development`.

(I noticed the need for this change because I got some data monitor
emails while I had the app running locally in what I thought was
development mode.)